backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
langgraph==0.2.60
langchain-core==0.3.83
langchain-groq==0.2.4
langchain-openai==0.2.14
langchain-anthropic==0.3.12
elevenlabs==2.37.0
pydub==0.25.1
python-dotenv==1.0.1
pyyaml==6.0.2
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
